move remote_map implementation to its own directory
